Homogeneous vectors
