+++ src/lib.rs
@@ -68,7 +68,8 @@ pub use scale::TypedScale;
 pub use transform2d::{Transform2D, TypedTransform2D};
 pub use transform3d::{Transform3D, TypedTransform3D};
 pub use point::{Point2D, Point3D, TypedPoint2D, TypedPoint3D, point2, point3};
-pub use vector::{HomogeneousVector, TypedVector2D, TypedVector3D, Vector2D, Vector3D, vec2, vec3};
+pub use vector::{TypedVector2D, TypedVector3D, Vector2D, Vector3D, vec2, vec3};
+pub use homogen::HomogeneousVector;
 
 pub use rect::{rect, Rect, TypedRect};
 pub use rotation::{Angle, Rotation2D, Rotation3D, TypedRotation2D, TypedRotation3D};
@@ -76,19 +77,21 @@ pub use side_offsets::{SideOffsets2D, TypedSideOffsets2D};
 pub use size::{Size2D, TypedSize2D, size2};
 pub use trig::Trig;
 
-pub mod approxeq;
-pub mod num;
-mod length;
 #[macro_use]
 mod macros;
-mod transform2d;
-mod transform3d;
+
+pub mod approxeq;
+mod homogen;
+pub mod num;
+mod length;
 mod point;
 mod rect;
 mod rotation;
 mod scale;
 mod side_offsets;
 mod size;
+mod transform2d;
+mod transform3d;
 mod trig;
 mod vector;
